# backend/docs/swagger.yaml
definitions:
...
  domain.App:
    properties:
      created_at:
        type: string
      id:
        type: string
      kb_id:
        type: string
      link:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/domain.AppSettings'
      type:
        $ref: '#/definitions/domain.AppType'
      updated_at:
        type: string
    type: object
  domain.AppListItem:
    properties:
      id:
        type: string
      link:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/domain.AppSettings'
      stats:
        $ref: '#/definitions/domain.ConversationStatForApp'
      type:
        $ref: '#/definitions/domain.AppType'
    type: object
  domain.AppSettings:
    properties:
      desc:
        type: string
      dingbot_aes_key:
        type: string
      dingbot_client_id:
        description: dingbot
        type: string
      dingbot_token:
        type: string
      dingbot_welcome_str:
        type: string
      feishubot_app_id:
        description: feishu
        type: string
      feishubot_app_secret:
        type: string
      feishubot_encrypt_key:
        type: string
      feishubot_verification_token:
        type: string
      feishubot_welcome_str:
        type: string
      icon:
        type: string
      recommend_doc_ids:
        items:
          type: string
        type: array
      recommend_questions:
        items:
          type: string
        type: array
      search_placeholder:
        type: string
      wecombot_agent_id:
        description: Wecom
        type: integer
      wecombot_corp_id:
        type: string
      wecombot_corp_secret:
        type: string
      wecombot_suite_encoding_aes_key:
        type: string
      wecombot_suite_id:
        type: string
      wecombot_suite_token:
        type: string
      wecombot_welcome_str:
        type: string
      welcome_str:
        type: string
    type: object
  domain.AppType:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-varnames:
    - AppTypeWeb
    - AppTypeWidget
    - AppTypeDingbot
    - AppTypeFeishu
    - AppTypeWecom
  domain.ChatRequest:
    properties:
      conversation_id:
        type: string
      message:
        type: string
      nonce:
        type: string
    required:
    - message
    type: object
  domain.ChatToKBMessage:
    properties:
      content:
        type: string
      role:
        type: string
    required:
    - content
    - role
    type: object
  domain.ChatToKBReq:
    properties:
      kb_id:
        type: string
      messages:
        items:
          $ref: '#/definitions/domain.ChatToKBMessage'
        type: array
    required:
    - kb_id
    - messages
    type: object
...
  domain.ConversationCount:
    properties:
      app_id:
        type: string
      count:
        type: integer
      date:
        type: string
    type: object
...
  domain.ConversationStatForApp:
    properties:
      app_id:
        type: string
      day_counts:
        items:
          $ref: '#/definitions/domain.ConversationCount'
        type: array
      last_24h_count:
        type: integer
      last_24h_ip_count:
        type: integer
    type: object
  domain.CreateAppReq:
    properties:
      icon:
        type: string
      kb_id:
        type: string
      name:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.AppType'
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
    required:
    - kb_id
    - type
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  domain.UpdateAppReq:
    properties:
      name:
        type: string
      settings:
        $ref: '#/definitions/domain.AppSettings'
    type: object
...
info:
  contact: {}
paths:
  /api/v1/app:
    delete:
      consumes:
      - application/json
      description: Delete app
      parameters:
      - description: app id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Delete app
      tags:
      - app
    post:
      consumes:
      - application/json
      description: Create app
      parameters:
      - description: create app request
        in: body
        name: create_app_request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateAppReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Create app
      tags:
      - app
    put:
      consumes:
      - application/json
      description: Update app
      parameters:
      - description: app
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Update app
      tags:
      - app
  /api/v1/app/detail:
    get:
      consumes:
      - application/json
      description: Get app detail
      parameters:
      - description: app id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.App'
              type: object
      summary: Get app detail
      tags:
      - app
  /api/v1/app/list:
    get:
      consumes:
      - application/json
      description: Get app list
      parameters:
      - description: kb id
        in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.AppListItem'
                  type: array
              type: object
      summary: Get app list
      tags:
      - app
  /api/v1/chat/kb:
    post:
      consumes:
      - application/json
      description: ChatToKB
      parameters:
      - description: ChatToKBReq Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.ChatToKBReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: ChatToKB
      tags:
      - chat
...
  /share/v1/app/link:
    get:
      consumes:
      - application/json
      description: Get app by link
      parameters:
      - description: link
        in: query
        name: link
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.App'
              type: object
      summary: Get app by link
      tags:
      - app
  /share/v1/chat/message:
    post:
      consumes:
      - application/json
      description: ChatMessage
      parameters:
      - description: app id
        in: query
        name: app_id
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: ChatMessage
      tags:
      - chat
swagger: "2.0"